+++ crates/l2/contracts/src/l1/interfaces/IOnChainProposer.sol
@@ -23,6 +23,10 @@ interface IOnChainProposer {
     /// @dev Event emitted when a batch is verified.
     event BatchVerified(uint256 indexed lastVerifiedBatch);
 
+    /// @notice A batch has been reverted.
+    /// @dev Event emitted when a batch is reverted.
+    event BatchReverted(bytes32 indexed newStateRoot);
+
     /// @notice Set the bridge address for the first time.
     /// @dev This method is separated from initialize because both the CommonBridge
     /// and the OnChainProposer need to know the address of the other. This solves
@@ -80,12 +84,19 @@ interface IOnChainProposer {
 
     /// @notice Method used to verify a batch of L2 blocks in Aligned.
     /// @param alignedPublicInputs The public inputs bytes of the proof.
-    /// @param alignedProgramVKey The public verifying key.
     /// @param alignedMerkleProof  The Merkle proof (sibling hashes) needed to reconstruct the Merkle root.
     function verifyBatchAligned(
         uint256 batchNumber,
         bytes calldata alignedPublicInputs,
-        bytes32 alignedProgramVKey,
         bytes32[] calldata alignedMerkleProof
     ) external;
+
+    /// @notice Allows unverified batches to be reverted
+    function revertBatch(uint256 batchNumber) external;
+
+    /// @notice Allows the owner to pause the contract
+    function pause() external;
+
+    /// @notice Allows the owner to unpause the contract
+    function unpause() external;
 }