+++ crates/networking/rpc/eth/transaction.rs
@@ -23,11 +23,7 @@ use ethrex_storage::Store;
 use ethrex_vm::{Evm, ExecutionResult};
 use serde::Serialize;
 
-#[cfg(feature = "l2")]
-use ethrex_common::types::Transaction;
 use serde_json::Value;
-#[cfg(feature = "l2")]
-use tracing::debug;
 use tracing::info;
 
 pub const ESTIMATE_ERROR_RATIO: f64 = 0.015;
@@ -174,7 +170,7 @@ impl RpcHandler for GetTransactionByBlockNumberAndIndexRequest {
             Some(block_number),
             block_header.hash(),
             Some(self.transaction_index),
-        );
+        )?;
         serde_json::to_value(tx).map_err(|error| RpcErr::Internal(error.to_string()))
     }
 }
@@ -221,7 +217,7 @@ impl RpcHandler for GetTransactionByBlockHashAndIndexRequest {
             Some(block_number),
             self.block,
             Some(self.transaction_index),
-        );
+        )?;
         serde_json::to_value(tx).map_err(|error| RpcErr::Internal(error.to_string()))
     }
 }
@@ -268,7 +264,7 @@ impl RpcHandler for GetTransactionByHashRequest {
             Some(block_number),
             block_hash,
             Some(index as usize),
-        );
+        )?;
         serde_json::to_value(transaction).map_err(|error| RpcErr::Internal(error.to_string()))
     }
 }
@@ -609,26 +605,13 @@ impl RpcHandler for SendRawTransactionRequest {
 
     async fn handle(&self, context: RpcApiContext) -> Result<Value, RpcErr> {
         let hash = if let SendRawTransactionRequest::EIP4844(wrapped_blob_tx) = self {
-            #[cfg(feature = "l2")]
-            {
-                debug!(
-                    "EIP-4844 transaction are not supported in the L2: {:#x}",
-                    Transaction::EIP4844Transaction(wrapped_blob_tx.tx.clone()).compute_hash()
-                );
-                return Err(RpcErr::InvalidEthrexL2Message(
-                    "EIP-4844 transactions are not supported in the L2".to_string(),
-                ));
-            }
-            #[cfg(not(feature = "l2"))]
-            {
-                context
-                    .blockchain
-                    .add_blob_transaction_to_pool(
-                        wrapped_blob_tx.tx.clone(),
-                        wrapped_blob_tx.blobs_bundle.clone(),
-                    )
-                    .await
-            }
+            context
+                .blockchain
+                .add_blob_transaction_to_pool(
+                    wrapped_blob_tx.tx.clone(),
+                    wrapped_blob_tx.blobs_bundle.clone(),
+                )
+                .await
         } else {
             context
                 .blockchain