+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Perf
 
+### 2025-06-27
+
+- Improve U256 decoding and PUSHX [#3332](https://github.com/lambdaclass/ethrex/pull/3332)
+
+### 2025-06-26
+
+- Refactor jump opcodes to use a blacklist on invalid targets.
+
+### 2025-06-20
+
+- Use a lookup table for opcode parsing [#3253](https://github.com/lambdaclass/ethrex/pull/3253)
+- Use specialized PUSH1 and PUSH2 implementations [#3262](https://github.com/lambdaclass/ethrex/pull/3262)
+
 ### 2025-05-27
 
 - Improved the performance of shift instructions. [2933](https://github.com/lambdaclass/ethrex/pull/2933)